Client side message inspector
This client side message inspector gets the user token, which is make
available to the application using a ShareObjects class, and appends it
to all outgoing service requests using a custom header. In the custom
header, TokenHeader, the user's session token is added. When the request
is received on the server side by the server message inspector in the
WCF123 Project, the XML is parsed for the Session token, the Service,
Service Call, and the Body.