Reflection!
In the case where there are service calls that do not need to be logged,
there are a number of ways to approach the situation. You could throw
the service calls into a giant if statement, but a more elegant method
is to create a class whose members are the values of the service call.
Afterwards, you can implement reflection in a class method to return a
list of the values of all properties in the class.