Reflection!
In some cases you may not want to log every service call that's being
made. There are several ways of dealing with this such as creating a
giant IF statment that checks to see if the service call you've
extracted from the XML file matches the service call you don't want to
log. A more elegant way of dealing with the situation is to create a
class that has as its' property values the name of the service calls.
Using reflection you can return a list of the 'exempt' service calls to
the message inspector, and search the list. This is a pretty SOLID and
DRY method of tackling this issue. See the ExemptServices class for a
quick rundown of how to use reflection