Designing Service Applications
A service is a public interface that provides access to a unit of
functionality. Services literally provide some programmatic service to
the caller, who consumes the service. Services are loosely coupled and
can be combined within a client, or combined within other services, to
provide functionality that is more complex. Services are distributable
and can be accessed from a remote machine as well as from the machine on
which they are running. Services are message-oriented, meaning that
service interfaces are defined by a Web Services Description Language
(WSDL) file, and operations are called using Extensible Markup Language
(XML)-based message schemas that are passed over a transport channel.
Services support a heterogeneous environment by focusing
interoperability at the message/interface definition. If components can
understand the message and interface definition, they can use the
service regardless of their base technology.