Restful Services With WCF
RESTful services are those which follow the REST (Representational State
Transfer) architectural style. Before implementing your first RESTful
service, lets first understand the concept behind it. As we know that
WCF allows us to make calls and exchange messages using SOAP over a
variety of protocols i.e. HTTP, TCP, Named Pipes and MSMQ etc. In a
scenario, if we are using SOAP over HTTP, we are just utilizing HTTP as
a transport. But HTTP is much more than just a transport. So, when we
talk about REST architectural style, it dictates that “Instead of using
complex mechanisms like CORBA, RPC or SOAP for communication, simply
HTTP should be used for making calls”.